requirements.txt
dlib~=19.24.0
ipyplot~=1.1.0
ipywidgets~=8.0.0
jupyterlab~=3.4.0
ml2rt~=0.2.0
Pillow~=9.1.0
pillow-heif~=0.7.0
protobuf~=3.20.0
python_dotenv~=0.21.0
redis~=4.3.0
redisai~=1.2.0
redis-om~=0.1.0
scikit-learn~=1.1.0
tqdm~=4.64.0
tritonclient[all]~=2.22.0